requirements.txt
pytest
black
isort

# remove_bg
rembg
pillow
